fix: middleware not working on Turbopack & auth enforcement
- Move middleware.ts → src/middleware.ts (Next.js 16 Turbopack ignores
  root middleware when using src/ directory)
- Protect API routes via middleware (previously all /api/ excluded)
- Remove JWT refresh interval caching — always fetch role/status from DB
- Add SessionProvider refetchInterval (30s) for near-realtime blocking
- Add AUTH_SECRET env var for NextAuth v5 compatibility
- Add .vercel to .gitignore

Closes #162
