+++ .gitignore
@@ -22,3 +22,4 @@ npm-debug.log*
 .DS_Store
 *.tsbuildinfo
 next-env.d.ts
+.vercel
