+++ src/middleware.ts
@@ -1,0 +1,61 @@
+import { NextResponse } from "next/server";
+import type { NextRequest } from "next/server";
+import { getToken } from "next-auth/jwt";
+
+export async function middleware(req: NextRequest) {
+  const { pathname } = req.nextUrl;
+  const isApiRoute = pathname.startsWith("/api/");
+
+  const token = await getToken({
+    req,
+    secret: process.env.AUTH_SECRET ?? process.env.NEXTAUTH_SECRET,
+    cookieName:
+      req.nextUrl.protocol === "https:"
+        ? "__Secure-authjs.session-token"
+        : "authjs.session-token",
+  });
+
+  // Not logged in
+  if (!token) {
+    if (isApiRoute) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
+    return NextResponse.redirect(new URL("/login", req.url));
+  }
+
+  const status = token.status as string | undefined;
+  const role = token.role as string | undefined;
+
+  // Rejected or pending
+  if (status === "rejected" || status === "pending") {
+    if (isApiRoute) {
+      return NextResponse.json({ error: "Forbidden" }, { status: 403 });
+    }
+    if (pathname !== "/pending") {
+      return NextResponse.redirect(new URL("/pending", req.url));
+    }
+  }
+
+  // Approved user on /pending → redirect to home
+  if (status === "approved" && pathname === "/pending") {
+    return NextResponse.redirect(new URL("/", req.url));
+  }
+
+  // Admin routes protection
+  if (pathname.startsWith("/admin") || pathname.startsWith("/api/admin")) {
+    if (role !== "admin") {
+      if (isApiRoute) {
+        return NextResponse.json({ error: "Forbidden" }, { status: 403 });
+      }
+      return NextResponse.redirect(new URL("/", req.url));
+    }
+  }
+
+  return NextResponse.next();
+}
+
+export const config = {
+  matcher: [
+    "/((?!api/auth|_next/static|_next/image|favicon.ico|login|pending).*)",
+  ],
+};
